Rename eip7594 to fulu, p1
