+++ .circleci/config.yml
@@ -155,6 +155,19 @@ jobs:
           command: make citest fork=electra
       - store_test_results:
           path: tests/core/pyspec/test-reports
+  test-fulu:
+    docker:
+      - image: cimg/python:3.12.4
+    working_directory: ~/specs-repo
+    steps:
+      - restore_cache:
+          key: v3-specs-repo-{{ .Branch }}-{{ .Revision }}
+      - restore_pyspec_cached_venv
+      - run:
+          name: Run py-tests
+          command: make citest fork=fulu
+      - store_test_results:
+          path: tests/core/pyspec/test-reports
   test-whisk:
     docker:
       - image: cimg/python:3.12.4
@@ -166,19 +179,6 @@ jobs:
       - run:
           name: Run py-tests
           command: make citest fork=whisk
-      - store_test_results:
-          path: tests/core/pyspec/test-reports
-  test-eip7594:
-    docker:
-      - image: cimg/python:3.12.4
-    working_directory: ~/specs-repo
-    steps:
-      - restore_cache:
-          key: v3-specs-repo-{{ .Branch }}-{{ .Revision }}
-      - restore_pyspec_cached_venv
-      - run:
-          name: Run py-tests
-          command: make citest fork=eip7594
       - store_test_results:
           path: tests/core/pyspec/test-reports
   table_of_contents:
@@ -239,10 +239,10 @@ workflows:
       - test-electra:
           requires:
             - install_pyspec_test
-      - test-whisk:
+      - test-fulu:
           requires:
             - install_pyspec_test
-      - test-eip7594:
+      - test-whisk:
           requires:
             - install_pyspec_test
       - table_of_contents
