+++ .github/workflows/run-tests.yml
@@ -86,7 +86,7 @@ jobs:
     needs: [lint,codespell,table_of_contents]
     strategy:
       matrix:
-        version: ["phase0", "altair", "bellatrix", "capella", "deneb", "electra", "whisk", "eip7594"]
+        version: ["phase0", "altair", "bellatrix", "capella", "deneb", "electra", "fulu", "whisk"]
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
