+++ pysetup/constants.py
@@ -5,7 +5,7 @@ BELLATRIX = 'bellatrix'
 CAPELLA = 'capella'
 DENEB = 'deneb'
 ELECTRA = 'electra'
-EIP7594 = 'eip7594'
+FULU = 'fulu'
 EIP6800 = 'eip6800'
 WHISK = 'whisk'
 EIP7732 = 'eip7732'
