+++ tests/core/pyspec/eth2spec/test/eip7594/unittests/test_config_invariants.py
@@ -1,11 +1,11 @@
 from eth2spec.test.context import (
     single_phase,
     spec_test,
-    with_eip7594_and_later,
+    with_fulu_and_later,
 )
 
 
-@with_eip7594_and_later
+@with_fulu_and_later
 @spec_test
 @single_phase
 def test_invariants(spec):
@@ -20,7 +20,7 @@ def test_invariants(spec):
     )
 
 
-@with_eip7594_and_later
+@with_fulu_and_later
 @spec_test
 @single_phase
 def test_polynomical_commitments_sampling(spec):
