+++ tests/generators/networking/main.py
@@ -1,14 +1,14 @@
 
-from eth2spec.test.helpers.constants import EIP7594
+from eth2spec.test.helpers.constants import FULU
 from eth2spec.gen_helpers.gen_from_tests.gen import run_state_test_generators, check_mods
 
 
 if __name__ == "__main__":
-    eip7594_mods = {key: 'eth2spec.test.eip7594.networking.test_' + key for key in [
+    fulu_mods = {key: 'eth2spec.test.fulu.networking.test_' + key for key in [
         'get_custody_columns',
     ]}
     all_mods = {
-        EIP7594: eip7594_mods
+        FULU: fulu_mods
     }
     check_mods(all_mods, "networking")
